# Conflicts:
#	conf/chesapeake_cvpr.yaml
#	environment.yml
#	evaluate.py
#	torchgeo/datamodules/__init__.py
#	torchgeo/samplers/single.py
#	torchgeo/trainers/byol.py
#	torchgeo/trainers/segmentation.py
#	train.py

+++ tests/models/test_changestar.py
@@ -18,7 +18,7 @@ BACKBONE = ["resnet18", "resnet34", "resnet50", "resnet101"]
 IN_CHANNELS = [64, 128]
 INNNR_CHANNELS = [16, 32, 64]
 NC = [1, 2, 4]
-SF = [4.0, 8.0, 1.0]
+SF = [4, 8, 1]
 
 
 class TestChangeStar:
@@ -65,7 +65,7 @@ class TestChangeStar:
         "inc,innerc,nc,sf", list(itertools.product(IN_CHANNELS, INNNR_CHANNELS, NC, SF))
     )
     def test_changemixin_output_size(
-        self, inc: int, innerc: int, nc: int, sf: float
+        self, inc: int, innerc: int, nc: int, sf: int
     ) -> None:
         m = ChangeMixin(
             in_channels=inc, inner_channels=innerc, num_convs=nc, scale_factor=sf
@@ -73,13 +73,13 @@ class TestChangeStar:
 
         y = m(torch.rand(3, 2, inc // 2, 32, 32))
         assert y[0].shape == y[1].shape
-        assert y[0].shape == (3, 1, int(32 * sf), int(32 * sf))
+        assert y[0].shape == (3, 1, 32 * sf, 32 * sf)
 
     @torch.no_grad()  # type: ignore[misc]
     def test_changestar(self) -> None:
         dense_feature_extractor = nn.modules.Sequential(
             nn.modules.Conv2d(3, 32, 3, 1, 1),
-            nn.modules.BatchNorm2d(32),  # type: ignore[no-untyped-call]
+            nn.modules.BatchNorm2d(32),
             nn.modules.ReLU(),
             nn.modules.MaxPool2d(3, 2, 1),
         )
@@ -106,7 +106,7 @@ class TestChangeStar:
     def test_changestar_invalid_inference_mode(self) -> None:
         dense_feature_extractor = nn.modules.Sequential(
             nn.modules.Conv2d(3, 32, 3, 1, 1),
-            nn.modules.BatchNorm2d(32),  # type: ignore[no-untyped-call]
+            nn.modules.BatchNorm2d(32),
             nn.modules.ReLU(),
             nn.modules.MaxPool2d(3, 2, 1),
         )
@@ -132,7 +132,7 @@ class TestChangeStar:
     def test_changestar_inference_output_size(self, inference_mode: str) -> None:
         dense_feature_extractor = nn.modules.Sequential(
             nn.modules.Conv2d(3, 32, 3, 1, 1),
-            nn.modules.BatchNorm2d(32),  # type: ignore[no-untyped-call]
+            nn.modules.BatchNorm2d(32),
             nn.modules.ReLU(),
             nn.modules.MaxPool2d(3, 2, 1),
         )
